Added admin functions

+++ server.js
@@ -18,6 +18,7 @@ app.use(express.json())
 app.post('/signup/',
 body('userId').isString().trim().escape(),
 body('password').isLength({ min: 8 }),
+body('admin').optional().isBoolean(),
 async function(req, res) {
 
    const errors = validationResult(req);
@@ -26,17 +27,23 @@ async function(req, res) {
    }
    const userId = req.body.userId
    const password = req.body.password
+   var admin = req.body.admin
+   if(admin == null){
+      admin = 0
+   }
 
-   const passwordHash = crypto.createHash('sha256').update(password).digest('hex')
+   const saltedPassword = userId + "+" + password
+   const passwordHash = crypto.createHash('sha256').update(saltedPassword).digest('hex')
    const userExists = await db.get(`SELECT * FROM users WHERE userID = (?)`, userId)
    if(userExists == null){
       const result = await db.run(`
       INSERT INTO users (userID, hashPwd, admin)
       VALUES (?,?,?);
-      `, userId, passwordHash, 0)
+      `, userId, passwordHash, admin)
       res.status(201).send({
          "userId":userId
-      });
+      })
+      return
    }
 
    res.status(403).send()
@@ -58,9 +65,11 @@ async function(req, res){
    if(user == null){
       //user doens't exist
       res.status(401).send()
+      return
    }
    const passwordHash = user.hashPwd
-   const signinPwdHash = crypto.createHash('sha256').update(password).digest('hex')
+   const saltedPassword = userId + "+" + password
+   const signinPwdHash = crypto.createHash('sha256').update(saltedPassword).digest('hex')
    if(passwordHash == signinPwdHash){
 
       const payload = {
@@ -96,10 +105,21 @@ app.use(async function(req, res, next) {
    next()
 })
 
+//log track of user behaviour
+app.use(async function(req, res, next) {
+   const userId = getUserIdFromToken(req.headers.authorization)
+   await db.run(`
+      INSERT INTO userlogs (userID, method, path, body)
+      VALUES (?,?,?,?);
+      `, userId, req.method, req.path, JSON.stringify(req.body))
+
+   next()
+})
+
 app.get('/hashMap/', async function (req, res) {
    const userId = getUserIdFromToken(req.headers.authorization)
    var hash = getUserHashMap(userId)
-   res.send(hash.getAll());
+   res.send(hash.getAll())
 })
 
 app.get('/hashMap/:key', function (req, res) {
@@ -108,12 +128,12 @@ app.get('/hashMap/:key', function (req, res) {
    var value = hash.get(req.params.key)
    res.send({
       "value":value
-   });
+   })
 })
 
 app.put('/hashMap/:key',
 body('value').isString().trim().escape(),
-function(req, res) {
+async function(req, res) {
 
    const errors = validationResult(req);
    if (!errors.isEmpty()) {
@@ -121,16 +141,76 @@ function(req, res) {
    }
    const userId = getUserIdFromToken(req.headers.authorization)
    var hash = getUserHashMap(userId)
-   var entry = hash.add(req.params.key, req.body.value)
+   var entry = await hash.add(req.params.key, req.body.value)
    res.status(201).send(entry);
-});
+})
 
 app.delete('/hashMap/:key', function(req, res) {
    const userId = getUserIdFromToken(req.headers.authorization)
    var hash = getUserHashMap(userId)
-   var array = hash.deleteValue(req.params.key)
+   hash.deleteValue(req.params.key)
    res.status(200).send()
-});
+})
+
+// ADMIN FUNCTIONS -------
+app.get('/user/', async function (req, res) {
+   const currentUserId = getUserIdFromToken(req.headers.authorization)
+   const admin = await isUserAdmin(currentUserId)
+   if(admin == 1){
+      const data = await db.all(`SELECT userID FROM users;`)
+      var users = []
+      data.forEach(userObject => {
+         users.push(userObject.userID)
+      })
+      res.status(200).send(users)
+   }
+
+   res.status(401).send()
+
+})
+
+app.get('/user/:userId/hashMap/', async function (req, res) {
+   const currentUserId = getUserIdFromToken(req.headers.authorization)
+   const admin = await isUserAdmin(currentUserId)
+   if(admin == 1){
+      const userId = req.params.userId
+      var hash = getUserHashMap(userId)
+      res.send(hash.getAll())
+      return
+   }
+
+   res.status(401).send()
+
+})
+
+app.get('/user/:userId/hashMap/:key', async function (req, res) {
+   const currentUserId = getUserIdFromToken(req.headers.authorization)
+   const admin = await isUserAdmin(currentUserId)
+   if(admin == 1){
+      const userId = req.params.userId
+      var hash = getUserHashMap(userId)
+      var value = hash.get(req.params.key)
+      res.send({
+         "value":value
+      })
+      return
+   }
+
+   res.status(401).send()
+})
+
+app.delete('/user/:userId/hashMap/:key', async function(req, res) {
+   const currentUserId = getUserIdFromToken(req.headers.authorization)
+   const admin = await isUserAdmin(currentUserId)
+   if(admin == 1){
+      const userId = req.params.userId
+      var hash = getUserHashMap(userId)
+      hash.deleteValue(req.params.key)
+      res.status(200).send()
+      return
+   }
+   res.status(401).send()
+})
 
 var server = app.listen(8082, async function () {
    var host = server.address().address
@@ -158,7 +238,7 @@ function getUserHashMap(userId){
 }
 
 async function loadHashMaps(){
-   const data = await db.all(`SELECT * FROM hashmap_table`)
+   const data = await db.all(`SELECT * FROM hashmap_table;`)
    data.forEach(datapoint => {
       const hash = getUserHashMap(datapoint.userID)
       hash.add(datapoint.hashKey, datapoint.value)
@@ -189,3 +269,11 @@ function getUserIdFromToken(token){
    return decoded.userId
 }
 
+async function isUserAdmin(userId){
+   const data = await db.get(`
+   SELECT admin
+   FROM users
+   WHERE userID = ?;`, userId)
+
+   return data.admin
+}
